Map Rendering + Saving bug fix + Better headers
Included more headers in comment files
Added rendering of map
Fixed bug writing map to file

+++ src/entity/include/map.h
@@ -4,7 +4,7 @@
 #include <SDL.h>
 
 typedef struct Map_Tile {
-    Uint32 texture_index;
+    Uint16 texture_index;
     Vector2D position;
 } Map_Tile;
